[Fix] Use EditorWindow.GetWindowWithRect and specify minSize for popup
There was a bug where the window was showing up in the wrong size on
some machines. This specifies a rect size for the window in its spawning
function, and specifies a minimum size for the window so that it is not
resizable and contains all of the content at all times.