Optimize scrollrect performance
Removes layouting components and does the layout calculations once in
the code instead of every scroll frame. Changes UIMask For UIRectMask2D
which is significantly faster at masking. Also disables the "pixel
perfect" option on the UI canvas, which introduces a very significant
performance increase when scrolling without a perceptible difference in
quality.