Adds dev configuration file for shopify employees
This adds a dev.yml and a gemfile which helps shopify employees
configure their environment faster. It is not required and wont affect
anything if you are not a shopify employee, but the gemfile might
actually help you if you use rbenv or chruby with some shim that
automatically switches what ruby you are using based on the gemfile.