[UI Toolkit] Add Class for verifying credentials in the editor
This class allows verifying credentials in the editor. It takes a
context object that implement IShopCredentials and is able to make
requests and mutate the verification state on that context object.

It also includes a delegate function for drawing the UI for verifying
credentials, it's designed to be initialized in OnEnable and called out
to in OnInspectorGUI