Allow injectable LoaderProvider for changing loader
Allows configuration of the loader used in the theme controllers via a
loader provider that instantiates a loader given the access token and
shop domain. This allows tests to use mock loaders while maintaining the
functionality of the client lazy load + cache. It also reworks some of
the cache invalidation in the controller to simplify the workflow, as
changing the loader now also invalidates the cached client.