[UI Toolkit] Adds a Stack-based view switcher + example
Allows pushing new views and maintaining a stack of views so that
navigating "back" always does the expected action. This will be used to
manage cart/product list/product details views in the Unity SDK factory
shops, and is designed to be easy to understand so that a 3rd party dev
can use it to build their own custom shops.